+++ crates/store/src/db/mod.rs
@@ -12,6 +12,7 @@ use miden_objects::{
     block::{Block, BlockNoteIndex},
     crypto::{hash::rpo::RpoDigest, merkle::MerklePath, utils::Deserializable},
     notes::{NoteId, NoteMetadata, Nullifier},
+    transaction::TransactionId,
     utils::Serializable,
     BlockHeader, GENESIS_BLOCK,
 };
@@ -48,6 +49,13 @@ pub struct NullifierInfo {
     pub block_num: BlockNumber,
 }
 
+#[derive(Debug, PartialEq)]
+pub struct TransactionSummary {
+    pub account_id: AccountId,
+    pub block_num: BlockNumber,
+    pub transaction_id: TransactionId,
+}
+
 #[derive(Debug, Clone, PartialEq)]
 pub struct NoteRecord {
     pub block_num: BlockNumber,
@@ -77,7 +85,7 @@ pub struct StateSyncUpdate {
     pub block_header: BlockHeader,
     pub chain_tip: BlockNumber,
     pub account_updates: Vec<AccountSummary>,
-    pub transactions: Vec<RpoDigest>,
+    pub transactions: Vec<TransactionSummary>,
     pub nullifiers: Vec<NullifierInfo>,
 }